+++ lib/idkit_extension.dart
@@ -1,4 +1,5 @@
 library idkit_extension;
 
 export 'src/list_extension.dart';
+export 'src/num_extension.dart';
 export 'src/string_extension.dart';