+++ test/idkit_extension_test.dart
@@ -2,6 +2,6 @@ import 'package:idkit_extension/idkit_extension.dart';
 import 'package:idkit_extension/src/list_extension.dart';
 
 void main() {
-  String a = '6666001222'.thousands();
+  String a = '6' + '0' * (-3 * -1);
   print(a);
 }